Validate timezone input in group-by date queries
Adds IANA timezone allowlist validation for the timezone parameter used
in group-by date expressions. Moves the IANA_TIME_ZONES constant to
twenty-shared so it can be reused across packages.
